Api/requirements.txt
imutils==0.5.4
matplotlib==3.7.1
numpy==1.24.3
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
pandas==2.0.1
Pillow==9.5.0
Pillow==9.5.0
tensorflow==2.10.1
tqdm==4.65.0
absl-py==1.4.0
astunparse==1.6.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dm-tree==0.1.8
docopt==0.6.2
Flask==2.3.2
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.4.0
gensim==4.3.1
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mediapipe==0.9.3.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pipreqs==0.4.13
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.0.9
pytz==2023.3
pywin32==305
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
smart-open==6.3.0
sounddevice==0.4.6
stellargraph==1.2.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-model-optimization==0.7.4
termcolor==2.3.0
threadpoolctl==3.1.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
Werkzeug==2.3.3
wrapt==1.15.0
yarg==0.1.9
zipp==3.15.0